chore: move TileProvider in their own folder
